refactor: the types were moved to another file
